Script to for player interaction
The player can touch whatever they want and we'll get the feedback in the Debug.log. Set the cube as "Kiwi" to see for point system setup but since the player can touch everything, we can set if else statements for what happens when they touch a certain item. Need to look into a scoreboard setup as well.